+++ users/models.py
@@ -7,6 +7,7 @@ from django.contrib.auth.signals import user_logged_in
 from phone_field import PhoneField
 from rlcis.models import IndustryType
 
+
 """
 Ref:https://www.codingforentrepreneurs.com/blog/how-to-create-a-custom-django-user-model/
 https://youtu.be/Ae7nc1EGv-A
@@ -84,19 +85,20 @@ class Users(AbstractBaseUser, PermissionsMixin):
     def __str__(self):
         return self.user_name
 
-    # """
-    # Create signal for first time login
-    # https://stackoverflow.com/questions/49385582/can-i-check-if-a-user-is-logged-in-for-the-first-time-after-this-user-is-logged
-    # More info on signals explained in https://youtu.be/FdVuKt_iuSI?list=PL-osiE80TeTtoQCKZ03TU5fNfx2UY6U4p
-    # """    
-    # def update_first_login(sender, user, *args, **kwargs):
-    #     if user.last_login is None:
-    #         # First time this user has logged in
-    #         kwargs['request'].session['first_login'] = True #Add first login attribute to session
-    # # Update the last_login value as normal
-    #     update_last_login(sender, user, **kwargs)
+    """
+    Create signal for first time login
+    https://stackoverflow.com/questions/49385582/can-i-check-if-a-user-is-logged-in-for-the-first-time-after-this-user-is-logged
+    More info on signals explained in https://youtu.be/FdVuKt_iuSI?list=PL-osiE80TeTtoQCKZ03TU5fNfx2UY6U4p
+    """   
+     
+    def update_first_login(sender, user, *args, **kwargs):
+        if user.last_login is None:
+            # First time this user has logged in
+            kwargs['request'].session['first_login'] = True #Add first login attribute to session
+    # Update the last_login value as normal
+        update_last_login(sender, user, **kwargs)
 
-    # user_logged_in.disconnect(update_last_login)
-    # user_logged_in.connect(update_first_login)
+    user_logged_in.disconnect(update_last_login)
+    user_logged_in.connect(update_first_login)
