--- users/templates/users/login-submit.html
@@ -1,60 +1,0 @@
-{% extends "rlcis/base.html" %}
-{% load static %}
-{% load crispy_forms_tags %}
-{% block content %}
-
-<html>
-    <head>
-  
-        <link rel="stylesheet" type="text/css" href="{% static 'users/style.css' %}"/>
-    </head>
-<body>
-
-<div class="container container-table">
-    <div class="row pt-3"></div>
-
- 
-    <div class="d-flex justify-content-center">
-        <div class="col-xs-12 col-sm-6 col-sm-offset-2 profile-main-div">
-            <hr>
-            <div class="row d-flex justify-content-center">
-                <p class="head-msg col-sm-offset-3">RLCIS Login </p>
-                <div class="col-xs-2 col-xs-offset-5 text-center column-div">
-                </div>
-            </div>  
-            <form class="form-horizontal profile-sub-form" method=POST>
-                <div class="row">
-                    <div class="col-sm-8">
-                    {% csrf_token %}
-                    {{ form.errors }}
-                    </div>
-                </div>
-                <div class="row d-flex justify-content-center">
-                    <div class="col-sm-8">
-                        <div class="form-group">
-                            {{ form|crispy }}
-                        </div>
-                        <div>
-                            <div class="border-top pt-3"></div>
-                        </div>
-                        <small class="float-right py-2 showPasswordToggle">SHOW</small>
-                    </div>
-                </div>    
-                <div class="col-sm-6">
-                    <button type="submit" class="btn btn-success accept-btn pull-right">Login</button>                     
-                </div>         
-                    <div class='container'>
-                        <small class="text-muted">
-                            <span class="ml-0"></span> Need An Account?</span> <a href="{% url 'users:register' %}">Sign Up Now</a>
-                            <span class="ml-4"></span> Forgot Password? <a href="{% url 'users:password_reset' %}">Reset Password</a>
-                        </small>
-                    <div>              
-                </div>
-            </form>                       
-        </div>       
-    </div>
-</div>
-</body>
-</html>
-{% endblock content %}
-<script src="{% static 'js/register.js' %}"></script>
