+++ users/templates/users/password_reset.html
@@ -13,18 +13,6 @@
                    </div>                    
                 {% csrf_token %}
                 {% crispy form %}  
-                <!-- <div class="row d-flex justify-content-center message-row">
-                        <p class="head-msg col-sm-offset-3"></p>
-                        {% if messages %}
-                        {% for message in messages %}
-                            <div class="row justify-content-center">                                       
-                                    <div class="row alert alert-{{ message.tags }}" id="mess_con">
-                                        {{ message }}
-                                    </div>
-                            </div>
-                        {% endfor %}
-                        {% endif %}        
-                </div>  -->
                 <hr>             
             </div>            
         </div>
