--- users/templates/users/password_reset.save.html
@@ -1,32 +1,0 @@
-{% extends "rlcis/base.html" %}
-{% load static %}
-{% load crispy_forms_tags %}
-{% block content %}
-
-
-    <div class="content-section">
-        <div class="container-fluid" style="background-color:#CDCDCD; max-width:450px">
-            <form method="POST">
-                {% csrf_token %}
-                <fieldset class="form-group">
-                    <legend class="border-bottom mb-4">Reset Password</legend>
-                    {{ form|crispy }}
-                </fieldset>
-                <div class="form-group col-md-10 ">
-                    <input type="email"
-                      name="email"
-                      placeholder="email address"
-                      class="form-control form-control-sm"
-                      id="emailField" required
-                    />        
-                <br>
-                <div class="form-group">
-                    <button class="btn btn-info accept-btn" type="submit">Request Password Reset</button>
-                </div>
-            </form>
-        </div>
-        <hr color = "FFF">
-        <br>
-    </div>
-
-{% endblock content %}
