+++ users/templates/users/register.html
@@ -28,20 +28,7 @@
                         </div>
                     </div>                                                     
                 </div>
-                <!-- <div class="col-12">
-                <div class="row message-row justify-content-center">   
-                    {% if messages %}
-                    {% for message in messages %}
-                        <div class="row">                                       
-                                <div class="row alert alert-{{ message.tags }}" id="mess_con">
-                                    {{ message }}
-                                </div>
-                        </div>
-                    {% endfor %}
-                    {% endif %}   
-                </div>                      
-                </div>               -->
-            </form>
+             </form>
             <div class="border-top">
                 <small class="text-muted">
                     Already Have An Account? <a class="ml-2" href="{% url 'users:login' %}">Sign In</a>
