+++ users/templates/users/set-new-password.html
@@ -19,7 +19,6 @@
         color: #7a797e;
     }
 </style>
-
         <div class="d-flex justify-content-center">
             <div class="col-xs-12 col-sm-4 col-sm-offset-2 profile-main-div">
                     <hr>
