--- users/templates/users/set-new-password.save.html
@@ -1,59 +1,0 @@
-{% extends 'rlcis/base.html' %}
-{% load static %}
-
-
-{% block content %}
-
-<div class="row">
-  <div class="col-md-4"></div>
-  <div class="col-md-4">
-    <div class="card mt-5">
-      <div class="container-fluid" style="background-color:#CDCDCD; max-width:500px">
-        <div class="card-body" >
-           
-          <legend class="border-bottom mb-4">New Reset Password</legend>
-
-        <form action="{% url 'users:set-new-password' uidb64 token %}" method="post">
-          {% csrf_token %}    
-          New password*
-            <div class="form-group" >
-              <input 
-                type="password"
-                name="password"
-                placeholder="Password"
-                class="form-control form-control-sm"
-                id="passwordField" 
-              />
-              <small class="float-right py-2 showPasswordToggle">SHOW</small>
-            </div>
-          New password confirmation*
-            <div class="form-group">    
-              <input
-                type="password"
-                name="password2"
-                placeholder="Confirm Password"
-                class="form-control form-control-sm"
-                id="passwordField"
-              />
-              <small class="float-right py-2 showPasswordToggle">SHOW</small>
-            </div>
-          <input type="submit" value="Submit" class="btn btn-outline-info" />
-          <hr color = "white">
-        </form>
-          <div class="container">
-              <small class="text-muted">
-                <span class="ml-0"></span>Have An Account?</span> <a href="{% url 'users:login' %}">Sign Up Now</a>
-                <span class="ml-4"></span>No account? <a href="{% url 'users:register' %}">Register</a>
-          </div>
-        </small>
-      </div>
-
-      </div>
-    </div>
-  </div>
-  <div class="col-md-4"></div>
-</div>
-
-<script src="{% static 'js/register.js' %}"></script>
-
-{% endblock content %}
