--- users/templates/users/test.html
@@ -1,23 +1,0 @@
-{% extends "rlcis/base.html" %}
-{% load static %}
-{% load crispy_forms_tags %}
-{% block content %}
-
-<link rel="stylesheet" type="text/css" href="{% static 'users/style.css' %}"/>
-
-        <div class="d-flex justify-content-center">
-            <div class="col-xs-12 col-sm-4 col-sm-offset-2 profile-main-div">
-                    <hr>
-                    <div class="row d-flex justify-content-center">
-                        <p class="head-msg col-sm-offset-3">RLCIS Registration </p>
-                    </div> 
-                    {% csrf_token %}
-                    {% crispy form %}  
-                    <ul>
-                        <li>You password must contain at least 8 characters</li>
-                        <li>Your password can't be a commonly used password</li>
-                    </ul>
-            </div>            
-        </div>    
-
-{% endblock content %}
